requirements.txt
pandas
openpyxl
requests
tabulate
colored
python-dotenv